+++ services/append.js
@@ -76,12 +76,12 @@ var append = {
   },
 
   // append a field to an angular template form
-  formField: function (content, name) {
+  formField: function (content, tpl) {
     var newContent = content.toString();
     // http://regexr.com/3e2he
     return newContent.replace(
-      new RegExp('(\n\\W+<!-- Buttons --)', 'gm'),
-      name
+      new RegExp('\n(\\W+<!-- Buttons --)', 'gm'),
+      '\n\n' + tpl + '$1'
     );
   }
 };